requirements.txt
aiosqlite==0.17.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
bcrypt==3.2.0
boto3==1.22.0
botocore==1.25.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
databases==0.5.5
Deprecated==1.2.13
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.75.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
jmespath==1.0.0
limits==2.6.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pylogbeat==2.0.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-logstash-async==2.5.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing_extensions==4.1.1
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.2
wrapt==1.14.0
